perf(cache): wrap markdown API routes in "use cache" builds
Cache Sanity reads and markdown serialization for all 12 .md routes via
tag-invalidated build functions, switching the four remaining static
fetchers to Live sanityFetch inside those scopes.
