+++ src/actions/laboratory-fetch/sanity.ts
@@ -1,4 +1,4 @@
-import { sanityFetchStatic } from "@/service/sanity"
+import { sanityFetch, sanityFetchStatic } from "@/service/sanity"
 import { imageFragment } from "@/service/sanity/queries"
 
 export interface LabProject {
@@ -16,16 +16,26 @@ export interface LabProject {
   } | null
 }
 
+const labProjectsQuery = /* groq */ `*[_type == "labProject"] {
+  title,
+  url,
+  description,
+  cover ${imageFragment}
+}`
+
 export async function fetchLabProjects(): Promise<LabProject[]> {
   // Called from a server action (not a `"use cache"` scope), so use the non-Live
   // fetch — the Live `sanityFetch`'s `cacheTag()` throws outside `use cache`.
   return sanityFetchStatic<LabProject[]>({
-    query: /* groq */ `*[_type == "labProject"] {
-      title,
-      url,
-      description,
-      cover ${imageFragment}
-    }`,
+    query: labProjectsQuery,
+    perspective: "published"
+  })
+}
+
+/** For `/api/lab.md` — caller provides a `"use cache"` scope. */
+export async function fetchLabProjectsPublished(): Promise<LabProject[]> {
+  return sanityFetch<LabProject[]>({
+    query: labProjectsQuery,
     perspective: "published"
   })
 }
