+++ src/app/(site)/(plain)/(content)/careers/[slug]/sanity.ts
@@ -32,7 +32,7 @@ export interface CareerPosition {
 
 export async function fetchCareerPosition(
   slug: string,
-  /** Pass `published: true` for non-draft contexts (e.g. the `.md` endpoint) — disables stega so output isn't polluted with invisible chars. */
+  /** Pass `published: true` inside a `"use cache"` scope (e.g. the `.md` build) — pins published perspective so stega stays off and no dynamic APIs are touched. */
   options?: { published?: boolean }
 ): Promise<CareerPosition | null> {
   const query = /* groq */ `*[_type == "openPosition" && slug.current == $slug][0]{
@@ -52,9 +52,10 @@ export async function fetchCareerPosition(
     }
   }`
   if (options?.published) {
-    return sanityFetchStatic<CareerPosition | null>({
+    return sanityFetch<CareerPosition | null>({
       query,
-      params: { slug }
+      params: { slug },
+      perspective: "published"
     })
   }
   return sanityFetch<CareerPosition | null>({
