+++ src/app/api/careers/[slug]/route.ts
@@ -1,9 +1,9 @@
 import * as Sentry from "@sentry/nextjs"
 import { NextResponse } from "next/server"
 
-import { fetchCareerPosition } from "@/app/(site)/(plain)/(content)/careers/[slug]/sanity"
 import { SITE_URL } from "@/lib/constants"
-import { portableTextToMarkdown } from "@/service/sanity/portable-text-to-markdown"
+
+import { buildCareerMarkdown } from "./markdown"
 
 const MD_HEADERS = {
   "Content-Type": "text/markdown; charset=utf-8",
@@ -23,42 +23,9 @@ export async function GET(
   const slug = rawSlug.slice(0, -3)
 
   try {
-    const position = await fetchCareerPosition(slug, { published: true })
-    // Closed positions 404 here too, mirroring the HTML page's `notFound()`.
-    if (!position || !position.isOpen) {
-      return new NextResponse("# 404 Not Found\n", {
-        status: 404,
-        headers: MD_HEADERS
-      })
-    }
-
-    const skills = position.applyFormSetup?.skills
-      ?.map((s) => s.title)
-      .filter(Boolean)
-
-    const parts: Array<string | null> = [
-      `# ${position.title}`,
-      "",
-      position.type ? `**Type:** ${position.type}` : null,
-      position.employmentType
-        ? `**Employment Type:** ${position.employmentType}`
-        : null,
-      position.location ? `**Location:** ${position.location}` : null,
-      position.applyUrl ? `**Apply:** ${position.applyUrl}` : null,
-      skills?.length ? `**Skills:** ${skills.join(", ")}` : null,
-      "",
-      "---",
-      "",
-      portableTextToMarkdown(position.jobDescription, { baseUrl: SITE_URL }),
-      "",
-      "---",
-      "",
-      `[View all content](${SITE_URL}/sitemap.md)`
-    ]
-
-    const markdown = parts.filter((part) => part !== null).join("\n")
-
+    const { markdown, status } = await buildCareerMarkdown(slug)
     return new NextResponse(markdown, {
+      status,
       headers: {
         ...MD_HEADERS,
         Link: `<${SITE_URL}/careers/${slug}>; rel="canonical"`
