+++ src/app/api/faq.md/route.ts
@@ -1,8 +1,9 @@
 import * as Sentry from "@sentry/nextjs"
 import { NextResponse } from "next/server"
 
-import { fetchFaqPage } from "@/app/(site)/(plain)/(content)/faq/sanity"
 import { SITE_URL } from "@/lib/constants"
+
+import { buildFaqMarkdown } from "./markdown"
 
 const MD_HEADERS = {
   "Content-Type": "text/markdown; charset=utf-8",
@@ -12,38 +13,9 @@ const MD_HEADERS = {
 
 export async function GET() {
   try {
-    const faq = await fetchFaqPage({ published: true })
-
-    if (!faq?.entries.length) {
-      return new NextResponse("# 404 Not Found\n", {
-        status: 404,
-        headers: MD_HEADERS
-      })
-    }
-
-    const body = faq.entries.flatMap((entry) => [
-      `## ${entry.question}`,
-      "",
-      entry.answer,
-      ""
-    ])
-
-    const parts = [
-      `# ${faq.heading || "FAQ"}`,
-      "",
-      faq.intro,
-      faq.intro ? "" : null,
-      "---",
-      "",
-      ...body,
-      "---",
-      "",
-      `[View all content](${SITE_URL}/sitemap.md)`
-    ]
-
-    const markdown = parts.filter((part) => part !== null).join("\n")
-
+    const { markdown, status } = await buildFaqMarkdown()
     return new NextResponse(markdown, {
+      status,
       headers: {
         ...MD_HEADERS,
         Link: `<${SITE_URL}/faq>; rel="canonical"`
